backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Task Queue
celery==5.3.4
kombu==5.3.4
amqp==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# API & HTTP
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
python-dateutil==2.8.2

# Image Processing
pillow==10.1.0
python-magic==0.4.27

# Email & Notifications
emails==0.6
jinja2==3.1.2
aiosmtplib==3.0.1
sendgrid==6.11.0
twilio==8.10.0

# Payment Gateways
stripe==7.8.0
razorpay==1.4.1
razorpay==1.4.1

# Data Validation & Serialization
phonenumbers==8.13.26
email-validator==2.1.0
orjson==3.9.10

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Performance
cachetools==5.3.2
msgpack==1.0.7

# S3 Storage
boto3==1.34.0
botocore==1.34.0

# Excel/CSV Export
openpyxl==3.1.2
pandas==2.1.4

# Rate Limiting
slowapi==0.1.9

# WebSockets
websockets==12.0
